refactor(pii): group recognizers by region
Splits the flat `recognizers/` directory into `generic/` plus
`{ca,eu,uk,us}/` subgroups, mirroring upstream's
predefined_recognizers layout. Region scope is now visible from the
filesystem path. Each subgroup re-exports `Recognizer` so leaf files
keep their `use super::Recognizer;` import unchanged.

Public API at `dbmcp_pii::recognizers::*` is preserved via flat
re-exports in mod.rs. The `all()` registry order and behaviour are
byte-identical to before. all.rs is merged back into mod.rs.
