+++ crates/pii/src/recognizers/ca/mod.rs
@@ -1,0 +1,7 @@
+//! Canada-specific recognizers.
+
+pub(super) use super::Recognizer;
+
+mod sin_ca;
+
+pub use sin_ca::sin_ca;
