+++ crates/pii/src/recognizers/eu/mod.rs
@@ -1,0 +1,7 @@
+//! Pan-EU recognizers.
+
+pub(super) use super::Recognizer;
+
+mod vat_number;
+
+pub use vat_number::vat_number;
