+++ crates/pii/src/recognizers/us/mod.rs
@@ -1,0 +1,15 @@
+//! US-specific recognizers.
+
+pub(super) use super::Recognizer;
+
+mod itin;
+mod passport_us;
+mod routing_number_us;
+mod tax_id_ein;
+mod us_ssn;
+
+pub use itin::itin;
+pub use passport_us::passport_us;
+pub use routing_number_us::routing_number_us;
+pub use tax_id_ein::tax_id_ein;
+pub use us_ssn::us_ssn;
